+++ tests/test_cleanups.py
@@ -7,6 +7,17 @@ from slash.loader import Loader
 from .utils import TestCase
 
 
+def test_test_cleanups_happen_before_fixture_cleanups(populated_suite, suite_test):
+    fixture = populated_suite.add_fixture()
+    suite_test.add_fixture(fixture)
+    fixture_cleanup = fixture.add_cleanup()
+    test_cleanup = suite_test.add_cleanup()
+    results = populated_suite.run()
+    test_result = results[suite_test]
+    events = test_result.data['events']
+    assert events.index(fixture_cleanup.get_event_id()) > events.index(test_cleanup.get_event_id()) >= 0
+
+
 def test_cleanups(populated_suite, suite_test):
 
     suite_test.add_cleanup()